backend/requirements.txt
fastapi>=0.110,<0.116
uvicorn[standard]>=0.24,<0.30
python-dotenv>=1.0
requests>=2.31
python-multipart>=0.0.6

# Database + models
SQLAlchemy==1.4.52
psycopg2-binary>=2.9
pydantic==2.11.4
email-validator>=2.2
PyJWT==2.10.1

# Data + algorithms stack
numpy>=1.26,<2.3

pandas>=2.0,<2.3
scipy== 1.11.4
scikit-learn>=1.3,<1.5
joblib>=1.3,<2
matplotlib>=3.7,<3.10
filterpy>=1.4.5
networkx>=3.1,<4
shapely>=2.0,<3
geopandas>=0.14,<1
pyproj>=3.5,<4
fiona>=1.9.5
rtree>=1.0.1
osmnx>=1.3,<2
geopy>=2.3


